+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `ML99_GEN_SYM`
  - More statement chaining macros: `ML99_CHAIN_EXPR_STMT`, `ML99_INTRODUCE_NON_NULL_PTR_TO_STMT`.
  - The ability to specify several variables in `ML99_INTRODUCE_VAR_TO_STMT`.
+ - `ML99_DETECTOR` handling all the C11 keywords and the underscore character (`_`).
 
 ### Deprecated